mirror_repo/requirements.txt
# Core dependencies for Looker Migrator
jinja2>=3.1.0
pyyaml>=6.0
pybars3>=0.9.7

# For DAX API integration and online PBIT compilation
requests>=2.28
websocket-client>=1.5
